requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.2
python-dotenv>=1.0.0
r2pipe>=1.7.0
click>=8.1.7
rich>=13.7.0
loguru>=0.7.2
pydantic>=2.5.2
typer>=0.9.0
openai>=1.12.0
ollama>=0.1.6
networkx>=3.4.2